Split `SwiftSDKGenerator` into multiple files
`SwiftSDKGenerator.swift` has grown quite large even at this early stage of supporting only macOS as a host platform and a few Linux distributions as target platforms. It's time to split it up for better readability and easier navigation.

(cherry picked from commit 42663aed647aa7ce5b4fc84ed4b71d5b23f06d74)
